+++ src/openassay/__init__.py
@@ -8,7 +8,7 @@ from openassay.acceptance import AcceptanceResult, LevelStats, run_acceptance
 from openassay.backcalc import BackCalcResult, Sample, back_calculate, back_calculate_many
 from openassay.curve import CalibrationResult, StandardCurve, fit_standard_curve
 from openassay.ingest import read_plate
-from openassay.plate import PlateData, PlateLayout, PlateWell, Well
+from openassay.plate import CollapsedReplicate, PlateData, PlateLayout, PlateWell, Well
 from openassay.range import RangeResult, determine_lloq_uloq
 from openassay.report import report_run
 
@@ -21,6 +21,7 @@ __all__ = [
     "PlateData",
     "PlateLayout",
     "PlateWell",
+    "CollapsedReplicate",
     "Sample",
     "StandardCurve",
     "Well",